+++ source/common/types.F90
@@ -1026,7 +1026,7 @@ contains
     if (t_inc%newsosol .and. .not.t_inc%nosoc) nspin = 1
 
     ! Gather gmat so that all processors have the full matrix
-    ihelp = lmmaxd*lmmaxd*t_inc%nqdos*nspin ! *t_inc%IELAST*t_inc%NSPIN*natypd
+    ihelp = lmmaxd*lmmaxd*t_inc%nqdos*nspin ! *t_inc%IELAST*NSPIN*natypd
     if (t_mpi_c_grid%dims(1)>1) then
       recvcounts = ntot_pt*ihelp
       displs = ioff_pt*ihelp